pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Richard <richardfelipelives@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
